修正自动数据类型的运算，完善半精度

+++ Detector.py
@@ -50,8 +50,8 @@ class Probe:
         orderr:
             空间项阶数
         '''
-        self.coeff_pe = cp.asarray(coeff_pe)
-        self.coeff_time = cp.asarray(coeff_time)
+        self.coeff_pe = cp.asarray(coeff_pe, dtype=cp.float16)
+        self.coeff_time = cp.asarray(coeff_time, dtype=cp.float16)
         self.pmt_pos = pmt_pos
         self.ordert = max(self.coeff_pe.shape[0], self.coeff_time.shape[0])
         self.orderr = max(self.coeff_pe.shape[1], self.coeff_time.shape[1])
@@ -66,6 +66,7 @@ class Probe:
         rho = cp.clip(rho, 0, 1)
         # calculate cos theta
         cos_theta = cp.cos(cp.arctan2(cp.linalg.norm(cp.cross(v[:, None, :], self.pmt_pos), axis=-1), cp.dot(v, self.pmt_pos.T)))
+        # lpmv 会返回 float32
         base_t = lpmv(cp.zeros(self.ordert, dtype=cp.float16)[:, None, None], cp.arange(self.ordert, dtype=cp.int16)[:, None, None], cos_theta[None, :, :])
         base_r = lpmv(cp.zeros(self.orderr, dtype=cp.float16)[:, None], cp.arange(self.orderr, dtype=cp.int16)[:, None], rho[None, :])
         return base_t, base_r
@@ -82,9 +83,9 @@ class Probe:
         '''
         base_t, base_r = self.genBase(vertex)
         # 计算空间项
-        NPE = cp.exp(cp.sum(cp.matmul(base_t[:self.coeff_pe.shape[0]].T, self.coeff_pe) * base_r[:self.coeff_pe.shape[1]].T[None, :, :], axis=2))
+        NPE = cp.exp(cp.sum(cp.matmul(base_t[:self.coeff_pe.shape[0]].T, self.coeff_pe, dtype=cp.float16) * base_r[:self.coeff_pe.shape[1]].T[None, :, :], axis=2, dtype=cp.float16))
         # 计算 R
-        Ti = cp.sum(cp.matmul(base_t[:self.coeff_time.shape[0]].T, self.coeff_time) * base_r[:self.coeff_time.shape[1]].T[None, :, :], axis=2)
+        Ti = cp.sum(cp.matmul(base_t[:self.coeff_time.shape[0]].T, self.coeff_time, dtype=cp.float16) * base_r[:self.coeff_time.shape[1]].T[None, :, :], axis=2, dtype=cp.float16)
         t = PEt - Ti.T[:,:, None] - vertex[:, -1][:, None, None]
         R = tau * (1 - tau) / ts * cp.exp(- cp.where(t < 0, t * (tau - 1), t * tau) / ts) * NPE.T[:, :, None] / E0
         if not sum_mode:
@@ -119,9 +120,10 @@ class Probe:
         '''
         base_t, base_r = self.genBasech(vertex, chs)
         # 计算空间项
-        NPE = cp.exp(cp.sum((base_t[:self.coeff_pe.shape[0]].T @ self.coeff_pe) * base_r[:self.coeff_pe.shape[1]].T, axis=1))
+        # @ 和 dot 无法指定数据类型
+        NPE = cp.exp(cp.sum((base_t[:self.coeff_pe.shape[0]].T @ self.coeff_pe) * base_r[:self.coeff_pe.shape[1]].T, axis=1, dtype=cp.float16))
         # 计算 R
-        Ti = cp.sum((base_t[:self.coeff_time.shape[0]].T @ self.coeff_time) * base_r[:self.coeff_time.shape[1]].T, axis=1)
+        Ti = cp.sum((base_t[:self.coeff_time.shape[0]].T @ self.coeff_time) * base_r[:self.coeff_time.shape[1]].T, axis=1, dtype=cp.float16)
         t = PEt - Ti.T[:, None] - vertex[:, -1][:, None]
         R = tau * (1 - tau) / ts * cp.exp(- cp.where(t < 0, t * (tau - 1), t * tau) / ts) * NPE.T[:, None] * vertex[:, 3][:, None] / E0
         # 分类计算 R 的积分
